fix: Params set
