+++ app/[locale]/publication_international/page.tsx
@@ -9,9 +9,9 @@ type Props = {
   params: Promise<{ locale: string }>;
 };
 
-export default async function PublicationInternationalPage({
-  params: { locale },
-}: Props) {
+export default async function PublicationInternationalPage({ params }: Props) {
+  const { locale } = await params;
+  setRequestLocale(locale);
   const papers = await getPublicationData("international");
   const t = await getTranslations({ locale });
