+++ types/activity.ts
@@ -1,0 +1,9 @@
+export type Activity = {
+  title: string;
+  description: string;
+};
+
+export type ActivityInfo = {
+  year: string;
+  activity: Activity[];
+};
