+++ types/index.ts
@@ -1,0 +1,3 @@
+export * from './activity';
+export * from './publication';
+export * from './common';
